move middleware/logging/extractors.py > data_extractors.py
