+++ docs/lib/reference/data_extractors.rst
@@ -1,0 +1,6 @@
+data_extractors
+===============
+
+
+.. automodule:: starlite.data_extractors
+    :members:
