+++ docs/lib/reference/middleware/logging.rst
@@ -1,0 +1,5 @@
+logging
+=======
+
+.. automodule:: starlite.middleware.logging
+    :members:
