--- docs/lib/reference/middleware/logging/extractors.rst
@@ -1,6 +1,0 @@
-extractors
-==========
-
-
-.. automodule:: starlite.middleware.logging.extractors
-    :members:
