--- docs/lib/reference/middleware/logging/index.rst
@@ -1,7 +1,0 @@
-logging
-=======
-
-.. toctree::
-
-    middleware
-    extractors
