--- docs/lib/reference/middleware/logging/middleware.rst
@@ -1,5 +1,0 @@
-middleware
-==========
-
-.. automodule:: starlite.middleware.logging.middleware
-    :members:
