--- starlite/middleware/logging/__init__.py
@@ -1,3 +1,0 @@
-from .middleware import LoggingMiddleware, LoggingMiddlewareConfig
-
-__all__ = ("LoggingMiddleware", "LoggingMiddlewareConfig")
